requirements.txt
selenium~=4.0.0a5
pymongo~=3.10.1
pandas~=1.0.3
requests~=2.23.0
beautifulsoup4~=4.9.0
lxml~=4.5.0
python-dotenv~=0.13.0
dash~=1.11.0
plotly~=4.6.0